Sample rename and Readme Updates
