+++ src/MAUI/Maui.Samples/Samples/Map/SetBasemap/SetBasemap.xaml
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage x:Class="ArcGIS.Samples.ChangeBasemap.ChangeBasemap"
+<ContentPage x:Class="ArcGIS.Samples.SetBasemap.SetBasemap"
              xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:esriUI="clr-namespace:Esri.ArcGISRuntime.Maui;assembly=Esri.ArcGISRuntime.Maui"
