--- src/WPF/WPF.Viewer/Samples/Map/ChangeBasemap/readme.metadata.json
@@ -1,30 +1,0 @@
-{
-    "category": "Map",
-    "description": "Change a map's basemap. A basemap is beneath all layers on a `Map` and is used to provide visual reference for the operational layers.",
-    "formal_name": "ChangeBasemap",
-    "ignore": false,
-    "images": [
-        "ChangeBasemap.jpg"
-    ],
-    "keywords": [
-        "basemap",
-        "basemap gallery",
-        "map",
-        "toolkit"
-    ],
-    "offline_data": [],
-    "redirect_from": [
-        "/net/latest/wpf/sample-code/change-basemap.htm"
-    ],
-    "relevant_apis": [
-        "Basemap",
-        "BasemapGallery",
-        "Map",
-        "MapView"
-    ],
-    "snippets": [
-        "ChangeBasemap.xaml.cs",
-        "ChangeBasemap.xaml"
-    ],
-    "title": "Change basemap"
-}
