+++ src/WPF/WPF.Viewer/Samples/Map/SetBasemap/SetBasemap.jpg
