+++ src/WPF/WPF.Viewer/Samples/Map/SetBasemap/SetBasemap.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGIS.WPF.Samples.ChangeBasemap.ChangeBasemap"
+﻿<UserControl x:Class="ArcGIS.WPF.Samples.SetBasemap.SetBasemap"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:esri="http://schemas.esri.com/arcgis/runtime/2013">
