+++ src/WPF/WPF.Viewer/Samples/Map/SetBasemap/SetBasemap.xaml.cs
@@ -9,17 +9,17 @@
 
 using Esri.ArcGISRuntime.Mapping;
 
-namespace ArcGIS.WPF.Samples.ChangeBasemap
+namespace ArcGIS.WPF.Samples.SetBasemap
 {
     [ArcGIS.Samples.Shared.Attributes.Sample(
-        name: "Change basemap",
+        name: "Set basemap",
         category: "Map",
         description: "Change a map's basemap. A basemap is beneath all layers on a `Map` and is used to provide visual reference for the operational layers.",
         instructions: "When the basemap gallery appears, select a basemap to be displayed.",
         tags: new[] { "basemap", "basemap gallery", "map", "toolkit" })]
-    public partial class ChangeBasemap
+    public partial class SetBasemap
     {
-        public ChangeBasemap()
+        public SetBasemap()
         {
             InitializeComponent();
