--- src/WinUI/ArcGIS.WinUI.Viewer/Samples/Map/ChangeBasemap/readme.md
@@ -1,29 +1,0 @@
-# Change basemap
-
-Change a map's basemap. A basemap is beneath all layers on a `Map` and is used to provide visual reference for the operational layers.
-
-![Image of change basemap](ChangeBasemap.jpg)
-
-## Use case
-
-Basemaps should be selected contextually. For example, in maritime applications, it would be more appropriate to use a basemap of the world's oceans as opposed to a basemap of the world's streets.
-
-## How to use the sample
-
-Use the drop down menu to select the active basemap from the list of available basemaps.
-
-## How it works
-
-1. Create an `Map` object.
-2. Set the map to the `MapView` object.
-3. Choose a new basemap type and set it on the map.
-
-## Relevant API
-
-* Basemap
-* Map
-* MapView
-
-## Tags
-
-basemap, map
