--- src/WinUI/ArcGIS.WinUI.Viewer/Samples/Map/ChangeBasemap/readme.metadata.json
@@ -1,27 +1,0 @@
-{
-    "category": "Map",
-    "description": "Change a map's basemap. A basemap is beneath all layers on a `Map` and is used to provide visual reference for the operational layers.",
-    "formal_name": "ChangeBasemap",
-    "ignore": false,
-    "images": [
-        "ChangeBasemap.jpg"
-    ],
-    "keywords": [
-        "basemap",
-        "map"
-    ],
-    "offline_data": [],
-    "redirect_from": [
-        "/net/latest/winui/sample-code/change-basemap.htm"
-    ],
-    "relevant_apis": [
-        "Basemap",
-        "Map",
-        "MapView"
-    ],
-    "snippets": [
-        "ChangeBasemap.xaml.cs",
-        "ChangeBasemap.xaml"
-    ],
-    "title": "Change basemap"
-}
