+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Map/SetBasemap/SetBasemap.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="ArcGIS.WinUI.Samples.ChangeBasemap.ChangeBasemap"
+﻿<UserControl x:Class="ArcGIS.WinUI.Samples.SetBasemap.SetBasemap"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:esriUI="using:Esri.ArcGISRuntime.UI.Controls">
