+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Map/SetBasemap/readme.metadata.json
@@ -1,30 +1,30 @@
 {
     "category": "Map",
     "description": "Change a map's basemap. A basemap is beneath all layers on a `Map` and is used to provide visual reference for the operational layers.",
-    "formal_name": "ChangeBasemap",
+    "formal_name": "SetBasemap",
     "ignore": false,
     "images": [
-        "changebasemap.jpg"
+        "SetBasemap.jpg"
     ],
     "keywords": [
-        "basemap",
+        "basemap style",
         "basemap gallery",
         "map",
         "toolkit"
-    ],
+  ],
     "offline_data": [],
     "redirect_from": [
-        "/net/latest/maui/sample-code/change-basemap.htm"
+        "/net/latest/winui/sample-code/set-basemap.htm"
     ],
     "relevant_apis": [
-        "Basemap",
+        "Basemap.Style",
         "BasemapGallery",
         "Map",
         "MapView"
-    ],
+  ],
     "snippets": [
-        "ChangeBasemap.xaml.cs",
-        "ChangeBasemap.xaml"
+        "SetBasemap.xaml.cs",
+        "SetBasemap.xaml"
     ],
-    "title": "Change basemap"
+    "title": "Set basemap"
 }
